+++ CHANGELOG.md
@@ -1,7 +1,26 @@
 # Change Log
+
 All notable changes to the "slackpresence" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
-- Initial release
+## 0.0.1
+
+Initial release of slackpresence
+
+## 0.0.2
+
+Share your workspace name as well
+
+## 0.0.3
+
+You don't _have_ to share your workspace name
+
+## 0.0.4
+
+* Add Vue support
+* Bump tar from 2.2.1 to 2.2.2 to fix [CVE-2018-20834](https://nvd.nist.gov/vuln/detail/CVE-2018-20834)
+
+## 0.0.5
+
+* Rewrite in TypeScript